优化在线

+++ app/src/main/java/itor/topnetwork/com/dxditor/bean/GjxxBean.java
@@ -8,7 +8,7 @@ public class GjxxBean {
     private int typeCode;
     private String typeName;
     private int typeCount;
-    private int typeProportion;
+    private double typeProportion;
 
     public GjxxBean(int typeCode, String typeName, int typeCount, int typeProportion) {
         this.typeCode = typeCode;
@@ -41,11 +41,11 @@ public class GjxxBean {
         this.typeCount = typeCount;
     }
 
-    public int getTypeProportion() {
+    public double getTypeProportion() {
         return typeProportion;
     }
 
-    public void setTypeProportion(int typeProportion) {
+    public void setTypeProportion(double typeProportion) {
         this.typeProportion = typeProportion;
     }
 }
